lane detection 流程跑通了，目前支持detr模型

+++ lanedet/core/line/assigners/hungarian_assigner.py
@@ -136,6 +136,6 @@ class HungarianAssigner(BaseAssigner):
         assigned_gt_inds[:] = 0
         # assign foregrounds based on matching results
         assigned_gt_inds[matched_row_inds] = matched_col_inds + 1
-        assigned_labels[matched_row_inds] = gt_labels[matched_col_inds]
+        assigned_labels[matched_row_inds] = gt_labels[matched_col_inds].long()
         return AssignResult(
             num_gts, assigned_gt_inds, None, labels=assigned_labels)
